Initial VTK Output implementation -Adapted the code from the previous assignment to output 3D data sets -Hacked compute*() to just return 0, so we can have something to output -Fixed a bug in FINDEXOF, which was accidentally indexing WAY outside the range it should
I'm not 100% sure the VTK output is correct, yet, but ParaView can open the files without any errors. In particular, the size of our Density output needed to be smaller than the size of our Velocity output due to a check on the size of the data that ParaView makes when importing CELL_DATA (see lines 242-253 of this file: https://github.com/Kitware/VTK/blob/master/Common/DataModel/vtkStructuredGrid.h ). We were doing this last time, as well, but I'm not certain that we're doing it right for our data set this time since all of our velocity values will depend on a previously-computed density value.